+++ djvm/src/main/kotlin/net/corda/djvm/analysis/AnalysisContext.kt
@@ -1,10 +1,10 @@
 package net.corda.djvm.analysis
 
+import net.corda.djvm.code.asPackagePath
 import net.corda.djvm.messages.MessageCollection
 import net.corda.djvm.references.ClassHierarchy
 import net.corda.djvm.references.EntityReference
 import net.corda.djvm.references.ReferenceMap
-import net.corda.djvm.source.ClassSource
 
 /**
  * The context in which one or more classes are analysed.
@@ -13,13 +13,11 @@ import net.corda.djvm.source.ClassSource
  * @property classes List of class definitions that have been analyzed.
  * @property references A collection of all referenced members found during analysis together with the locations from
  * where each member has been accessed or invoked.
- * @property inputClasses The classes passed in for analysis.
  */
 class AnalysisContext private constructor(
         val messages: MessageCollection,
         val classes: ClassHierarchy,
-        val references: ReferenceMap,
-        val inputClasses: List<ClassSource>
+        val references: ReferenceMap
 ) {
 
     private val origins = mutableMapOf<String, MutableSet<EntityReference>>()
@@ -28,7 +26,7 @@ class AnalysisContext private constructor(
      * Record a class origin in the current analysis context.
      */
     fun recordClassOrigin(name: String, origin: EntityReference) {
-        origins.getOrPut(name.normalize()) { mutableSetOf() }.add(origin)
+        origins.getOrPut(name.asPackagePath) { mutableSetOf() }.add(origin)
     }
 
     /**
@@ -42,19 +40,13 @@ class AnalysisContext private constructor(
         /**
          * Create a new analysis context from provided configuration.
          */
-        fun fromConfiguration(configuration: AnalysisConfiguration, classes: List<ClassSource>): AnalysisContext {
+        fun fromConfiguration(configuration: AnalysisConfiguration): AnalysisContext {
             return AnalysisContext(
                     MessageCollection(configuration.minimumSeverityLevel, configuration.prefixFilters),
                     ClassHierarchy(configuration.classModule, configuration.memberModule),
-                    ReferenceMap(configuration.classModule),
-                    classes
+                    ReferenceMap(configuration.classModule)
             )
         }
-
-        /**
-         * Local extension method for normalizing a class name.
-         */
-        private fun String.normalize() = this.replace("/", ".")
 
     }