+++ djvm/src/test/kotlin/net/corda/djvm/analysis/ClassAndMemberVisitorTest.kt
@@ -21,7 +21,7 @@ class ClassAndMemberVisitorTest : TestBase() {
     @Test
     fun `can traverse classes`() {
         val classesVisited = mutableSetOf<ClassRepresentation>()
-        val visitor = object : ClassAndMemberVisitor() {
+        val visitor = object : ClassAndMemberVisitor(configuration, null) {
             override fun visitClass(clazz: ClassRepresentation): ClassRepresentation {
                 classesVisited.add(clazz)
                 return clazz
@@ -47,7 +47,7 @@ class ClassAndMemberVisitorTest : TestBase() {
     @Test
     fun `can traverse fields`() {
         val membersVisited = mutableSetOf<Member>()
-        val visitor = object : ClassAndMemberVisitor() {
+        val visitor = object : ClassAndMemberVisitor(configuration, null) {
             override fun visitField(clazz: ClassRepresentation, field: Member): Member {
                 membersVisited.add(field)
                 return field
@@ -77,7 +77,7 @@ class ClassAndMemberVisitorTest : TestBase() {
     @Test
     fun `can traverse methods`() {
         val membersVisited = mutableSetOf<Member>()
-        val visitor = object : ClassAndMemberVisitor() {
+        val visitor = object : ClassAndMemberVisitor(configuration, null) {
             override fun visitMethod(clazz: ClassRepresentation, method: Member): Member {
                 membersVisited.add(method)
                 return method
@@ -102,7 +102,7 @@ class ClassAndMemberVisitorTest : TestBase() {
     @Test
     fun `can traverse class annotations`() {
         val annotations = mutableSetOf<String>()
-        val visitor = object : ClassAndMemberVisitor() {
+        val visitor = object : ClassAndMemberVisitor(configuration, null) {
             override fun visitClassAnnotation(clazz: ClassRepresentation, descriptor: String) {
                 annotations.add(descriptor)
             }
@@ -118,9 +118,21 @@ class ClassAndMemberVisitorTest : TestBase() {
     private class TestClassWithAnnotations
 
     @Test
-    fun `can traverse member annotations`() {
+    fun `cannot traverse member annotations when reading`() {
         val annotations = mutableSetOf<String>()
-        val visitor = object : ClassAndMemberVisitor() {
+        val visitor = object : ClassAndMemberVisitor(configuration, null) {
+            override fun visitMemberAnnotation(clazz: ClassRepresentation, member: Member, descriptor: String) {
+                annotations.add("${member.memberName}:$descriptor")
+            }
+        }
+        visitor.analyze<TestClassWithMemberAnnotations>(context)
+        assertThat(annotations).isEmpty()
+    }
+
+    @Test
+    fun `can traverse member annotations when writing`() {
+        val annotations = mutableSetOf<String>()
+        val visitor = object : ClassAndMemberVisitor(configuration, Writer()) {
             override fun visitMemberAnnotation(clazz: ClassRepresentation, member: Member, descriptor: String) {
                 annotations.add("${member.memberName}:$descriptor")
             }
@@ -146,7 +158,7 @@ class ClassAndMemberVisitorTest : TestBase() {
     @Test
     fun `can traverse class sources`() {
         val sources = mutableSetOf<String>()
-        val visitor = object : ClassAndMemberVisitor() {
+        val visitor = object : ClassAndMemberVisitor(configuration, null) {
             override fun visitSource(clazz: ClassRepresentation, source: String) {
                 sources.add(source)
             }
@@ -160,9 +172,21 @@ class ClassAndMemberVisitorTest : TestBase() {
     }
 
     @Test
-    fun `can traverse instructions`() {
+    fun `does not traverse instructions when reading`() {
         val instructions = mutableSetOf<Pair<Member, Instruction>>()
-        val visitor = object : ClassAndMemberVisitor() {
+        val visitor = object : ClassAndMemberVisitor(configuration, null) {
+            override fun visitInstruction(method: Member, emitter: EmitterModule, instruction: Instruction) {
+                instructions.add(Pair(method, instruction))
+            }
+        }
+        visitor.analyze<TestClassWithCode>(context)
+        assertThat(instructions).isEmpty()
+    }
+
+    @Test
+    fun `can traverse instructions when writing`() {
+        val instructions = mutableSetOf<Pair<Member, Instruction>>()
+        val visitor = object : ClassAndMemberVisitor(configuration, Writer()) {
             override fun visitInstruction(method: Member, emitter: EmitterModule, instruction: Instruction) {
                 instructions.add(Pair(method, instruction))
             }