+++ node/src/test/kotlin/net/corda/node/internal/cordapp/JarScanningCordappLoaderTest.kt
@@ -2,7 +2,9 @@ package net.corda.node.internal.cordapp
 
 import co.paralleluniverse.fibers.Suspendable
 import net.corda.core.flows.*
+import net.corda.node.VersionInfo
 import net.corda.node.cordapp.CordappLoader
+import net.corda.nodeapi.internal.PLATFORM_VERSION
 import net.corda.testing.node.internal.cordappsForPackages
 import net.corda.testing.node.internal.getTimestampAsDirectoryName
 import net.corda.testing.node.internal.packageInDirectory
@@ -101,13 +103,70 @@ class JarScanningCordappLoaderTest {
     @Test
     fun `cordapp classloader can load cordapp classes`() {
         val isolatedJAR = JarScanningCordappLoaderTest::class.java.getResource("isolated.jar")!!
-        val loader = JarScanningCordappLoader.fromJarUrls(listOf(isolatedJAR))
+        val loader = JarScanningCordappLoader.fromJarUrls(listOf(isolatedJAR), VersionInfo.UNKNOWN)
 
         loader.appClassLoader.loadClass(isolatedContractId)
         loader.appClassLoader.loadClass(isolatedFlowName)
     }
 
-    private fun cordappLoaderForPackages(packages: Iterable<String>): CordappLoader {
+    @Test
+    fun `cordapp classloader sets target and min version to 1 if not specified`() {
+        val jar = JarScanningCordappLoaderTest::class.java.getResource("versions/no-min-or-target-version.jar")!!
+        val loader = JarScanningCordappLoader.fromJarUrls(listOf(jar), VersionInfo.UNKNOWN)
+        loader.cordapps.filter { !it.info.shortName.equals("corda-core") }.forEach {
+            assertThat(it.info.targetPlatformVersion).isEqualTo(1)
+            assertThat(it.info.minimumPlatformVersion).isEqualTo(1)
+        }
+    }
+
+    @Test
+    fun `cordapp classloader returns correct values for minPlatformVersion and targetVersion`() {
+        // load jar with min and target version in manifest
+        // make sure classloader extracts correct values
+        val jar = JarScanningCordappLoaderTest::class.java.getResource("versions/min-2-target-3.jar")!!
+        val loader = JarScanningCordappLoader.fromJarUrls(listOf(jar), VersionInfo.UNKNOWN)
+        // exclude the core cordapp
+        val cordapp = loader.cordapps.filter { it.cordappClasses.contains("net.corda.core.internal.cordapp.CordappImpl")}.single()
+        assertThat(cordapp.info.targetPlatformVersion).isEqualTo(3)
+        assertThat(cordapp.info.minimumPlatformVersion).isEqualTo(2)
+    }
+
+    @Test
+    fun `cordapp classloader sets target version to min version if target version is not specified`() {
+        // load jar with minVersion but not targetVersion in manifest
+        val jar = JarScanningCordappLoaderTest::class.java.getResource("versions/min-2-no-target.jar")!!
+        val loader = JarScanningCordappLoader.fromJarUrls(listOf(jar), VersionInfo.UNKNOWN)
+        // exclude the core cordapp
+        val cordapp = loader.cordapps.filter { it.cordappClasses.contains("net.corda.core.internal.cordapp.CordappImpl")}.single()
+        assertThat(cordapp.info.targetPlatformVersion).isEqualTo(2)
+        assertThat(cordapp.info.minimumPlatformVersion).isEqualTo(2)
+    }
+
+    @Test
+    fun `cordapp classloader does not load apps when their min platform version is greater than the platform version`() {
+        val jar = JarScanningCordappLoaderTest::class.java.getResource("versions/min-2-target-3.jar")!!
+        val loader = JarScanningCordappLoader.fromJarUrls(listOf(jar), VersionInfo.UNKNOWN.copy(platformVersion = 1))
+        // exclude the core cordapp
+        assertThat(loader.cordapps.size).isEqualTo(1)
+    }
+
+    @Test
+    fun `cordapp classloader does load apps when their min platform version is less than the platform version`() {
+        val jar = JarScanningCordappLoaderTest::class.java.getResource("versions/min-2-target-3.jar")!!
+        val loader = JarScanningCordappLoader.fromJarUrls(listOf(jar), VersionInfo.UNKNOWN.copy(platformVersion = 1000))
+        // exclude the core cordapp
+        assertThat(loader.cordapps.size).isEqualTo(2)
+    }
+
+    @Test
+    fun `cordapp classloader does load apps when their min platform version is equal to the platform version`() {
+        val jar = JarScanningCordappLoaderTest::class.java.getResource("versions/min-2-target-3.jar")!!
+        val loader = JarScanningCordappLoader.fromJarUrls(listOf(jar), VersionInfo.UNKNOWN.copy(platformVersion = 2))
+        // exclude the core cordapp
+        assertThat(loader.cordapps.size).isEqualTo(2)
+    }
+
+    private fun cordappLoaderForPackages(packages: Iterable<String>, versionInfo: VersionInfo = VersionInfo.UNKNOWN): CordappLoader {
 
         val cordapps = cordappsForPackages(packages)
         return testDirectory().let { directory ->